# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Companies API",
    "description": "API for managing companies and their details.",
    "version": "1.0",
    "contact": {
      "name": "Tomislav Kušek",
      "email": "tomislav.kusek@fer.hr"
    },
    "license": {
      "name": "Creative Commons Zero v1.0 Universal",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/"
    }
  },
  "paths": {
    "/companies": {
      "get": {
        "summary": "Get all companies",
        "description": "Retrieve all companies from the database.",
        "responses": {
          "200": {
            "description": "List of companies retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "companies": { "type": "array", "items": { "$ref": "#/components/schemas/Company" } }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No companies found."
          }
        }
      },
      "post": {
        "summary": "Add a new company",
        "description": "Add a new company to the database.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Company" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company added successfully."
          },
          "400": {
            "description": "Invalid input data."
          }
        }
      }
    },
    "/companies/name/{name}": {
      "get": {
        "summary": "Get company by name",
        "description": "Retrieve a specific company by its name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Name of the company to retrieve.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Company retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "company": { "$ref": "#/components/schemas/Company" }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Company not found."
          }
        }
      }
    },
    "/companies/founder/{name}": {
      "get": {
        "summary": "Get companies by founder",
        "description": "Retrieve all companies founded by a specific individual.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Name of the founder.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Companies retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Company" }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No companies found for the specified founder."
          }
        }
      }
    },
    "/companies/foundedBefore/{year}": {
      "get": {
        "summary": "Get companies founded before a year",
        "description": "Retrieve all companies founded before a specific year.",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "description": "Year to filter companies by.",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Companies retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Company" }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No companies found before the specified year."
          }
        }
      }
    },
    "/companies/{name}": {
      "delete": {
        "summary": "Delete a company",
        "description": "Deletes a company by its name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "The name of the company to delete.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Company deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Company not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Company": {
        "type": "object",
        "properties": {
          "company_id": { "type": "string" },
          "company_name": { "type": "string" },
          "founder": { "type": "string" },
          "established_year": { "type": "integer" },
          "headquarters": { "type": "string" },
          "industry": { "type": "string" },
          "number_of_employees": { "type": "integer" },
          "annual_revenue": { "type": "integer" },
          "website": { "type": "string" },
          "products": {
            "type": "array",
            "items": { "type": "string" }
          },
          "description": { "type": "string" }
        },
        "required": ["company_name", "founder", "established_year"]
      }
    }
  }
}